add supporting tools

+++ laravel/app/Actions/DocumentManagement/EditDocumentAction.php
@@ -49,28 +49,27 @@ class EditDocumentAction
             if (key_exists('file', $validatedData)){
                 $validatedData['file']=$validatedData['file']->getClientOriginalName();
             }
+            if(key_exists('shareholder', $validatedData) || key_exists('director', $validatedData)){
+                if($document->recipient_name == NULL && $document->reminder_date == NULL){
+                    $document->signers()->detach();
+                }
+                if ($validatedData['shareholder'][0] != NULL){
+                    $implodedShareholderIds=implode(',',$validatedData['shareholder']);
+                    $invite_emails_shareholder=explode(',',$implodedShareholderIds);
+                    for ($x = 0; $x <= count($invite_emails_shareholder) - 1; $x++) {
+                        $document->signers()->attach([1 =>['user_id'=> $invite_emails_shareholder[$x], 'user_type'=>'shareholder']]);
+                    }
+                }
+                if ($validatedData['director'][0] != NULL){
+                    $implodedDirectorIds=implode(',',$validatedData['director']);
+                    $invite_emails_id_director=explode(',',$implodedDirectorIds);
+                    for ($x = 0; $x <= count($invite_emails_id_director) - 1; $x++) {
+                        $document->signers()->attach([1 =>['user_id'=> $invite_emails_id_director[$x], 'user_type'=>'director']]);
+                    }
+                }
+            }
 
-            if($document->recipient_name == NULL && $document->reminder_date == NULL){
-                $document->signers()->detach();
-            }
-            if ($validatedData['shareholder'][0] != NULL){
-                $implodedShareholderIds=implode(',',$validatedData['shareholder']);
-                $invite_emails_shareholder=explode(',',$implodedShareholderIds);
-                for ($x = 0; $x <= count($invite_emails_shareholder) - 1; $x++) {
-                    $document->signers()->attach([1 =>['user_id'=> $invite_emails_shareholder[$x], 'user_type'=>'shareholder']]);
-                }
-            }
-            if ($validatedData['director'][0] != NULL){
-                $implodedDirectorIds=implode(',',$validatedData['director']);
-                $invite_emails_id_director=explode(',',$implodedDirectorIds);
-                for ($x = 0; $x <= count($invite_emails_id_director) - 1; $x++) {
-                    $document->signers()->attach([1 =>['user_id'=> $invite_emails_id_director[$x], 'user_type'=>'director']]);
-                }
-            }
-//            if ($document->invited_at != null){
-//                (new CancelInviteAction())->execute($validatedData['document_hashed']);
-//                $validatedData['invited_at']=null;
-//            }
+
             $document->update($validatedData);
 
             DB::commit();
